refactor: Config 한 클래스만 있는 디렉토리를 config 디렉토리로 응집
